backend/requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
numpy==2.1.1
Pillow==11.3.0
opencv-python==4.10.0.84
